Refactors PR analysis checklist integration
